+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+**October 11, 2023 [DESCW-1452](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1452)**
+
+- Frontend
+- Backend
+  - fixed contracts section missing columns
+
 **October 6, 2023 [DESCW-1517](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1517)**
 
 - Frontend
 - Backend
   - Added update functionality to project closeout.
-
 
 **October 5, 2023 [DESCW-1513](https://apps.itsm.gov.bc.ca/jira/browse/DESCW-1513)**